+++ src/stories/mock/components/codes/analitic.js
@@ -1,0 +1,8 @@
+const def = ` <Analytics
+  idTracking={'UA-0000000-0'}
+  urlPath={'/'}
+  debug
+/>
+`;
+
+export default def;