+++ src/stories/mock/components/codes/backers-open-collective.js
@@ -1,0 +1,8 @@
+const def = `<BackersOpenCollective
+    imageRounded
+    imageBordered
+    sortDirection={'desc'}
+    collective={'webpack'}
+/>`;
+
+export default def;