+++ src/stories/mock/components/codes/browser.js
@@ -1,0 +1,8 @@
+const def = `<Browser
+  showBrowserVendor
+  showBrowserVersion
+  vendor={'chrome'}
+  version={'>= 20.0'}
+/>`;
+
+export default def;