+++ src/stories/mock/components/codes/gif.js
@@ -1,0 +1,11 @@
+const def = ` <Gif
+    image={'images/gif/write.gif'}
+    preview={'images/demo/develop.png'}
+/>`;
+const autoplay = ` <Gif
+    image={'images/gif/write.gif'}
+    preview={'images/demo/develop.png'}
+    autoplay
+/>`;
+
+export { def, autoplay };