+++ src/stories/mock/components/codes/image.js
@@ -1,0 +1,9 @@
+const def = `<Image
+  alt={'This is an image example'}
+  src={'images/backers/woman.svg'}
+  tooltip={'This is a woman'}
+  width={'200'}
+/>
+  `;
+
+export default def;