+++ src/stories/mock/components/codes/imagelist.js
@@ -1,0 +1,33 @@
+const def = `const companies = [
+    {
+      image: 'images/image-list/react.png'
+    },
+    {
+      image: 'images/image-list/graphql.png'
+    },
+    {
+      image: 'images/image-list/apollo.png'
+    },
+    {
+      image: 'images/image-list/storybook.png'
+    },
+    {
+      image: 'images/image-list/angularjs.png'
+    },
+    {
+      image: 'images/image-list/bootstrap.png'
+    },
+    {
+      image: 'images/image-list/browserstack.png'
+    },
+    {
+      image: 'images/image-list/jest.png'
+    }
+  ];
+  
+  <ImageList
+    gray
+    images={companies}
+  />`;
+
+export default def;