+++ src/stories/mock/components/codes/input.js
@@ -1,0 +1,12 @@
+const def = `<Input
+className={''}
+type={'text'}
+name={'myInput'}
+value={'This is an input'}
+placeholder={'Enter a value'}
+label={'Input label'}
+icon={'fa fa-user'}
+iconAlign={'left'}
+/>`;
+
+export default def;