+++ src/stories/mock/components/codes/languageswitcher.js
@@ -1,0 +1,18 @@
+const def = `<LanguageSwitcher
+  customLabels={{
+    DE: 'Deutsch',
+    ES: 'Español',
+    FR: 'Français',
+    US: 'English'
+  }}
+  languages={[
+    'US',
+    'FR',
+    'ES',
+    'DE'
+  ]}
+  placeholder="Select language..."
+  showLabel
+/>`;
+
+export default def;