+++ src/stories/mock/components/codes/link.js
@@ -1,0 +1,12 @@
+const def = `<Link
+  href={'https://front10.com'}
+  className={''}
+  tooltip={'This is a tooltip'}
+  target={''}
+>
+  <Icon icon="fa fa-home" />
+  <span>This is a link</span>
+</Link>
+  `;
+
+export default def;