+++ src/stories/mock/components/codes/sponsors.js
@@ -1,0 +1,26 @@
+const def = `const sponsors = [
+  {
+    image: 'images/sponsors/facebook.png'
+  },
+  {
+    image: 'images/sponsors/google.png'
+  },
+  {
+    image: 'images/sponsors/twitter.png'
+  },
+  {
+    image: 'images/sponsors/salesforce.png'
+  },
+  {
+    image: 'images/sponsors/amazon.png'
+  }
+];
+
+<Sponsors
+    ccsClassName={'px-3'}
+    gray
+    sponsors={sponsors}
+/>
+  `;
+
+export default def;