+++ changelog.md
@@ -9,12 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
+- [?] Prop `outline` to Button, `imageShadow`, `shadow`, `showFooter`, `outlineButton` to Features, by [@ycgarrido](https://github.com/ycgarrido)
 - [#149](https://github.com/front10/landing-page-book/pull/149) Props `shadow` to Image, `imageShadow`, `summaryJustified`, `shadow` to Card, by [@ycgarrido](https://github.com/ycgarrido)
 - [#149](https://github.com/front10/landing-page-book/pull/149) Theming variable `--Image-boxShadow`, `--Card-boxShadow`, by [@ycgarrido](https://github.com/ycgarrido)
 
 #### Change
 
 #### Remove
+
+- [?] Prop `showBorder` to Features, by [@ycgarrido](https://github.com/ycgarrido)
 
 ## [1.1.0] 2018-10-18